+++ leod/fmm_callers.py
@@ -8,16 +8,18 @@ Prepare and call the fast marching routine for a given precaluculated graph and
 input start/end points.
 """
 
+from numpy import array
 from math import pi
 from leod.fmm_polar_graph import find_vertex_index
 from leod.fmm_fast_marching import fast_marching
 from leod.fmm_fast_marching import endpoint_distance
+from leod.triangulation import find_closest_vertex
 
 def calculate_pair_distance(shape, vertex, start_point, end_point, order, graph_type="tri", grid=-1, is_radians=False):
     
     start_point_temp = [0.0, 0.0]
     end_point_temp = [0.0, 0.0]
-    # Convert to radians if input points given in degrees (assumed by default)
+    # Convert to radians if input points given in degrees (assumed by default).
     if is_radians == False:
         conv = pi / 180.0
     else:
@@ -25,17 +27,21 @@ def calculate_pair_distance(shape, vertex, start_point, end_point, order, graph_
     for i in range(2):
         start_point_temp[i] = start_point[i] * conv
         end_point_temp[i] = end_point[i] * conv
-            
+    
+    # Get Cartesian coordinatess of start and end points.
+    start_carts = shape.polar2cart(start_point_temp[0], start_point_temp[1])   
+    end_carts = shape.polar2cart(end_point_temp[0], end_point_temp[1])
+     
     # Find start and end vertices
     if graph_type == "polar":
         start_vertex, start_th_index, st_ph_index = find_vertex_index(grid.theta_list, grid.phi_list, start_point_temp[0], start_point_temp[1])
         end_vertex, end_th_index, end_ph_index = find_vertex_index(grid.theta_list, grid.phi_list, end_point_temp[0], end_point_temp[1])
-        start_carts = shape.polar2cart(start_point_temp[0], start_point_temp[1])
     else:
         # Use triangulation (to be implemented!)
-        x=0
+        start_vertex = find_closest_vertex(vertex, array(start_carts))
+        end_vertex = find_closest_vertex(vertex, array(end_carts))
     
-    end_carts = shape.polar2cart(end_point_temp[0], end_point_temp[1])
+    
     end_vertex_carts = vertex[end_vertex].carts
     
     # If closest vertex and endpoint are the same, no interpolation needed.